+++ tests/Remoting.Server/Startup.fs
@@ -21,15 +21,20 @@
 namespace Bolero.Tests.Remoting
 
 open System
+open System.Text.Json.Serialization
+open Bolero.Tests.Remoting.Client
 open Microsoft.AspNetCore
 open Microsoft.AspNetCore.Authentication.Cookies
 open Microsoft.AspNetCore.Builder
 open Microsoft.AspNetCore.Hosting
+open Microsoft.AspNetCore.Http
+open Microsoft.AspNetCore.Mvc
 open Microsoft.Extensions.DependencyInjection
 open Microsoft.Extensions.Hosting
 open Microsoft.Extensions.Logging
 open Bolero.Remoting.Server
 open Bolero.Server
+open FSharp.SystemTextJson.Swagger
 
 module Page =
     open Bolero.Html
@@ -93,26 +98,32 @@ type Startup() =
             .AddCookie()
             |> ignore
         services
-            .AddRemoting<MyApiHandler>()
+            .AddBoleroRemoting<MyApiHandler>()
             .AddBoleroHost()
             .AddServerSideBlazor()
         |> ignore
+        services.AddSwaggerForSystemTextJson(JsonFSharpOptions()) |> ignore
+        services.AddEndpointsApiExplorer() |> ignore
 
-    member this.Configure(app: IApplicationBuilder, env: IHostEnvironment) =
+    member this.Configure(app: IApplicationBuilder, env: IHostEnvironment, log: ILogger<Startup>) =
         app.UseAuthentication()
-            .UseRemoting()
             .UseStaticFiles()
+            .UseSwagger()
+            .UseSwaggerUI()
             .UseRouting()
+            .UseAuthorization()
             .UseBlazorFrameworkFiles()
             .UseEndpoints(fun endpoints ->
                 endpoints.MapBlazorHub() |> ignore
+                endpoints.MapBoleroRemoting()
+                    .WithOpenApi()
+                |> ignore
                 endpoints.MapFallbackToBolero(Page.index) |> ignore)
         |> ignore
 
         if env.IsDevelopment() then
             app.UseDeveloperExceptionPage()
                 .UseWebAssemblyDebugging()
-                |> ignore
 
 module Main =
     [<EntryPoint>]