+++ tests/Unit.Client/Startup.fs
@@ -41,7 +41,7 @@ module Program =
     let Main args =
         let builder = WebAssemblyHostBuilder.CreateDefault(args)
         builder.RootComponents.Add<Tests>("#app")
-        builder.Services.AddRemoting(builder.HostEnvironment) |> ignore
+        builder.Services.AddBoleroRemoting(builder.HostEnvironment) |> ignore
         builder.Services.AddScoped<AuthenticationStateProvider, DummyAuthProvider>() |> ignore
         builder.Services.AddAuthorizationCore() |> ignore
         builder.Build().RunAsync() |> ignore